Break up non-unique signature tests to work around apparent IntelliJ bug
