+++ core-plugin/testData/inspections/restSignatureInspection/nonUniqueRestSignaturesPostFunction2/expected.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Copyright (C) 2014 The Android Open Source Project
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~      http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+<problems>
+  <problem>
+    <file>test.java</file>
+    <line>38</line>
+    <description>Multiple methods with same rest path &quot;POST function2/{}&quot;: &quot;function2&quot; and &quot;function2&quot;</description>
+  </problem>
+</problems>
