+++ core-plugin/testData/inspections/restSignatureInspection/nonUniqueRestSignaturesPutCollectionResponse/src/test.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright (C) 2014 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.example.app;
+
+import com.google.api.server.spi.config.Api;
+import com.google.api.server.spi.config.ApiMethod;
+import com.google.api.server.spi.config.DefaultValue;
+import com.google.api.server.spi.config.Named;
+import com.google.api.server.spi.response.CollectionResponse;
+
+import java.lang.Boolean;
+import java.lang.String;
+import java.util.ArrayList;
+import java.util.List;
+
+
+@Api
+public class MyClass {
+  // "PUT collectionresponse_foo"
+  @ApiMethod(path="", httpMethod = "")
+  public CollectionResponse<Foo> update2(@DefaultValue @Named("id") String id){
+    return  null;
+  }
+
+  // "PUT collectionresponse_foo"
+  @ApiMethod(path="", httpMethod = "")
+  public CollectionResponse<Foo> update2(@DefaultValue @Named("id") int id){
+    return  null;
+  }
+
+  public class Foo{}
+}
