+++ core-plugin/testSrc/com/google/gct/idea/appengine/validation/RestSignatureInspectionTest.java
@@ -166,7 +166,41 @@ public class RestSignatureInspectionTest extends EndpointTestBase {
     Assert.assertEquals("boo", path);
   }
 
-  public void fixme_testNonUniqueRestSignatures() {
+  public void testNonUniqueRestSignaturesPostBoo() {
+    doTest();
+  }
+
+  public void testNonUniqueRestSignaturesPostFunction2() {
+    doTest();
+  }
+
+  public void testNonUniqueRestSignaturesPutCollectionResponse() {
+    doTest();
+  }
+
+  public void testNonUniqueRestSignaturesGet1() {
+    doTest();
+  }
+
+  // todo(elharo): this one may be a bug; research
+  public void fixme_testNonUniqueRestSignaturesGetFoo() {
+    doTest();
+  }
+
+  // todo(elharo): this one may be a bug; research
+  public void fixme_testNonUniqueRestSignaturesGetFooCollection() {
+    doTest();
+  }
+
+  public void testNonUniqueRestSignaturesGetFunction1() {
+    doTest();
+  }
+
+  public void testNonUniqueRestSignaturesGetList2() {
+    doTest();
+  }
+
+  public void testNonUniqueRestSignaturesGetList3() {
     doTest();
   }
 
@@ -182,7 +216,8 @@ public class RestSignatureInspectionTest extends EndpointTestBase {
     LocalInspectionTool localInspectionTool = new RestSignatureInspection();
     String testName = getTestName(true);
     myFixture.setTestDataPath(getTestDataPath());
-    myFixture.testInspection("inspections/restSignatureInspection/" + testName, new LocalInspectionToolWrapper(localInspectionTool));
+    LocalInspectionToolWrapper wrapper = new LocalInspectionToolWrapper(localInspectionTool);
+    myFixture.testInspection("inspections/restSignatureInspection/" + testName, wrapper);
   }
 
   private void initializePsiMethod(String methodName, String httpMethodValue, String pathValue) {
